Update to move delete output button to context menu
